backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
